Simplify files
