+++ cpp/dolfinx/io/CMakeLists.txt
@@ -9,7 +9,6 @@ set(HEADERS_io
   ${CMAKE_CURRENT_SOURCE_DIR}/xdmf_function.h
   ${CMAKE_CURRENT_SOURCE_DIR}/xdmf_mesh.h
   ${CMAKE_CURRENT_SOURCE_DIR}/xdmf_meshtags.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/xdmf_read.h
   ${CMAKE_CURRENT_SOURCE_DIR}/xdmf_utils.h
   PARENT_SCOPE
 )
